Remove unnecessary console log outs and refactor imports.
